Fix integer overflow when dealing with big matrices

+++ parsec/data_dist/matrix/irregular_tiled_matrix.c
@@ -260,7 +260,7 @@ int get_tile_count(const irregular_tiled_matrix_desc_t *desc, int m, int n)
 {
     assert(0 <= m && m < desc->mt && 0 <= n && n < desc->nt);
     uint64_t res = desc->Mtiling[m] * desc->Ntiling[n];
-    uint64_t max = desc->m * desc->n;
+    uint64_t max = (uint64_t)desc->m * (uint64_t)desc->n;
     (void)res;
     (void)max;
     assert(0 < res && res <= max);
